requirements.txt
Pillow
numpy
pandas
google-cloud-vision
# Levenshtein
# beautifulsoup4
